refactor(protocol-util): move ProtocolFactory up one package
Moves the `ProtocolFactory` one record up to better indicate intent.
