+++ protocol-test-util/pom.xml
@@ -94,7 +94,6 @@
     <dependency>
       <groupId>org.jeasy</groupId>
       <artifactId>easy-random-core</artifactId>
-      <version>5.0.0</version>
     </dependency>
 
     <!-- test dependencies -->
