+++ protocol-test-util/src/main/java/io/camunda/zeebe/test/broker/protocol/EnumRandomizer.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.test.broker.protocol.record;
+package io.camunda.zeebe.test.broker.protocol;
 
 import java.util.Objects;
 import org.jeasy.random.randomizers.AbstractRandomizer;
